requirements.txt
chromadb==1.0.16
sentence-transformers==5.1.0
langchain==0.3.27
langgraph==0.6.4
langchain-community==0.3.27
rank-bm25==0.2.2
langchain-google-genai==2.1.9
langchain-chroma==0.2.5
fastapi==0.116.1
uvicorn==0.35.0
streamlit
aiocache
audio-recorder-streamlit
SpeechRecognition
pyaudio
jinja2
python-dotenv
